separate search form from result rendering
