+++ src/App.jsx
@@ -5,7 +5,8 @@ import "./style/Departures.css";
 import { connect } from "react-redux";
 import { t } from "./Translations.jsx";
 import { BrowserRouter as Router, Route, withRouter } from "react-router-dom";
-import Departures from "./Departures.jsx";
+// import Departures from "./Departures.jsx";
+import SearchBus from "./SearchBus.jsx";
 
 class UnconnectedNavigation extends Component {
   changeLngFr = () => {
@@ -51,7 +52,7 @@ class UnconnectedApp extends Component {
       <div>
         <Router>
           <Navigation />
-          <Route exact={true} path="/" component={Departures} />
+          <Route exact={true} path="/" component={SearchBus} />
         </Router>
       </div>
     );
